requirements.txt
j1939==0.1.0.dev1
python-can==4.5.0
typing_extensions
rich == 14.0.0